ASM Customer List Facade refactor
