+++ feature-libs/asm/components/asm-toggle-ui/asm-toggle-ui.component.ts
@@ -5,7 +5,8 @@
  */
 
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { AsmService, AsmUi } from '@spartacus/asm/core';
+import { AsmService } from '@spartacus/asm/core';
+import { AsmUi } from '@spartacus/asm/root';
 import { Subscription } from 'rxjs';
 
 @Component({
