+++ feature-libs/asm/components/customer-selection/customer-selection.component.spec.ts
@@ -8,7 +8,8 @@ import {
 } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
-import { AsmConfig, AsmService, CustomerSearchPage } from '@spartacus/asm/core';
+import { AsmConfig, AsmService } from '@spartacus/asm/core';
+import { CustomerSearchPage } from '@spartacus/asm/root';
 import { GlobalMessageService, I18nTestingModule, User } from '@spartacus/core';
 import {
   DirectionMode,
