+++ feature-libs/asm/core/connectors/converters.ts
@@ -7,7 +7,7 @@
 import { InjectionToken } from '@angular/core';
 import { CustomerListsPage } from '@spartacus/asm/root';
 import { Converter } from '@spartacus/core';
-import { CustomerSearchPage } from '../models/asm.models';
+import { CustomerSearchPage } from '../../root/model/asm.models';
 
 export const CUSTOMER_SEARCH_PAGE_NORMALIZER = new InjectionToken<
   Converter<any, CustomerSearchPage>
