+++ feature-libs/asm/core/facade/asm-customer-list.service.ts
@@ -5,10 +5,18 @@
  */
 
 import { Injectable } from '@angular/core';
+import { select, Store } from '@ngrx/store';
 import { AsmCustomerListFacade, CustomerListsPage } from '@spartacus/asm/root';
 import { Query, QueryService, QueryState } from '@spartacus/core';
 import { Observable } from 'rxjs';
+import {
+  CustomerSearchOptions,
+  CustomerSearchPage,
+} from '../../root/model/asm.models';
 import { AsmConnector } from '../connectors/asm.connector';
+import { AsmActions } from '../store/actions/index';
+import { StateWithAsm } from '../store/asm-state';
+import { AsmSelectors } from '../store/index';
 
 @Injectable()
 export class AsmCustomerListService implements AsmCustomerListFacade {
@@ -20,7 +28,8 @@ export class AsmCustomerListService implements AsmCustomerListFacade {
 
   constructor(
     protected queryService: QueryService,
-    protected asmConnector: AsmConnector
+    protected asmConnector: AsmConnector,
+    protected store: Store<StateWithAsm>
   ) {}
 
   getCustomerLists(): Observable<CustomerListsPage | undefined> {
@@ -29,5 +38,37 @@ export class AsmCustomerListService implements AsmCustomerListFacade {
 
   getCustomerListsState(): Observable<QueryState<CustomerListsPage>> {
     return this.customerListQuery$.getState();
+  }
+
+  /**
+   * Search for customers in a customer list
+   */
+  customerListCustomersSearch(options: CustomerSearchOptions): void {
+    this.store.dispatch(new AsmActions.CustomerListCustomersSearch(options));
+  }
+
+  /**
+   * Returns the customer search result data for a customer list
+   */
+  getCustomerListCustomersSearchResults(): Observable<CustomerSearchPage> {
+    return this.store.pipe(
+      select(AsmSelectors.getCustomerListCustomersSearchResults)
+    );
+  }
+
+  /**
+   * Returns the customer list customers search result loading status.
+   */
+  getCustomerListCustomersSearchResultsLoading(): Observable<boolean> {
+    return this.store.pipe(
+      select(AsmSelectors.getCustomerListCustomersSearchResultsLoading)
+    );
+  }
+
+  /**
+   * Reset the customer list customers search result data to the initial state.
+   */
+  customerListCustomersSearchReset(): void {
+    this.store.dispatch(new AsmActions.CustomerListCustomersSearchReset());
   }
 }
