+++ feature-libs/asm/core/facade/asm.service.ts
@@ -11,7 +11,7 @@ import {
   AsmUi,
   CustomerSearchOptions,
   CustomerSearchPage,
-} from '../models/asm.models';
+} from '../../root/model/asm.models';
 import { AsmActions } from '../store/actions/index';
 import { StateWithAsm } from '../store/asm-state';
 import { AsmSelectors } from '../store/index';
@@ -54,44 +54,12 @@ export class AsmService {
   }
 
   /**
-   * Search for customers in a customer list
-   */
-  customerListCustomersSearch(options: CustomerSearchOptions): void {
-    this.store.dispatch(new AsmActions.CustomerListCustomersSearch(options));
-  }
-
-  /**
-   * Returns the customer search result data for a customer list
-   */
-  getCustomerListCustomersSearchResults(): Observable<CustomerSearchPage> {
-    return this.store.pipe(
-      select(AsmSelectors.getCustomerListCustomersSearchResults)
-    );
-  }
-
-  /**
-   * Returns the customer list customers search result loading status.
-   */
-  getCustomerListCustomersSearchResultsLoading(): Observable<boolean> {
-    return this.store.pipe(
-      select(AsmSelectors.getCustomerListCustomersSearchResultsLoading)
-    );
-  }
-
-  /**
    * Returns the customer list customers search result error status.
    */
   getCustomerListCustomersSearchResultsError(): Observable<boolean> {
     return this.store.pipe(
       select(AsmSelectors.getCustomerListCustomersSearchResultsError)
     );
-  }
-
-  /**
-   * Reset the customer list customers search result data to the initial state.
-   */
-  customerListCustomersSearchReset(): void {
-    this.store.dispatch(new AsmActions.CustomerListCustomersSearchReset());
   }
 
   /**
