+++ feature-libs/asm/core/public_api.ts
@@ -8,7 +8,6 @@ export * from './asm-core.module';
 export * from './config/asm-config';
 export * from './connectors/index';
 export * from './facade/index';
-export * from './models/index';
 export * from './services/index';
 export * from './store/actions/index';
 export * from './store/asm-state';
