+++ feature-libs/asm/core/store/actions/customer.action.spec.ts
@@ -2,7 +2,7 @@ import { StateUtils, User } from '@spartacus/core';
 import {
   CustomerSearchOptions,
   CustomerSearchPage,
-} from '../../models/asm.models';
+} from '../../../root/model/asm.models';
 import {
   CUSTOMER_LIST_CUSTOMERS_SEARCH_DATA,
   CUSTOMER_SEARCH_DATA,
