+++ feature-libs/asm/core/store/actions/customer.action.ts
@@ -8,7 +8,7 @@ import { StateUtils } from '@spartacus/core';
 import {
   CustomerSearchOptions,
   CustomerSearchPage,
-} from '../../models/asm.models';
+} from '../../../root/model/asm.models';
 import {
   CUSTOMER_LIST_CUSTOMERS_SEARCH_DATA,
   CUSTOMER_SEARCH_DATA,
