+++ feature-libs/asm/core/store/asm-state.ts
@@ -5,7 +5,7 @@
  */
 
 import { StateUtils } from '@spartacus/core';
-import { AsmUi, CustomerSearchPage } from '../models/asm.models';
+import { AsmUi, CustomerSearchPage } from '../../root/model/asm.models';
 
 export const ASM_FEATURE = 'asm';
 export const CUSTOMER_SEARCH_DATA = '[asm] Customer search data';
