+++ feature-libs/asm/core/store/effects/customer.effect.spec.ts
@@ -2,8 +2,8 @@ import { TestBed } from '@angular/core/testing';
 import { provideMockActions } from '@ngrx/effects/testing';
 import { cold, hot } from 'jasmine-marbles';
 import { Observable, of } from 'rxjs';
+import { CustomerSearchPage } from '../../../root/model/asm.models';
 import { AsmConnector } from '../../connectors/asm.connector';
-import { CustomerSearchPage } from '../../models/asm.models';
 import { AsmActions } from '../actions/index';
 import { CustomerEffects } from './customer.effect';
