+++ feature-libs/asm/core/store/effects/customer.effect.ts
@@ -9,8 +9,8 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { normalizeHttpError } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
 import { catchError, map, switchMap } from 'rxjs/operators';
+import { CustomerSearchPage } from '../../../root/model/asm.models';
 import { AsmConnector } from '../../connectors/asm.connector';
-import { CustomerSearchPage } from '../../models/asm.models';
 import { AsmActions } from '../actions/index';
 
 @Injectable()
