+++ feature-libs/asm/core/store/selectors/customer-search.selectors.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { select, Store, StoreModule } from '@ngrx/store';
 import { StateUtils, User } from '@spartacus/core';
-import { CustomerSearchPage } from '../../models/asm.models';
+import { CustomerSearchPage } from '../../../root/model/asm.models';
 import { AsmActions } from '../actions';
 import { StateWithAsm } from '../asm-state';
 import * as fromReducers from '../reducers/index';
