+++ feature-libs/asm/core/store/selectors/customer-search.selectors.ts
@@ -6,7 +6,7 @@
 
 import { createSelector, MemoizedSelector } from '@ngrx/store';
 import { StateUtils } from '@spartacus/core';
-import { CustomerSearchPage } from '../../models/asm.models';
+import { CustomerSearchPage } from '../../../root/model/asm.models';
 import { AsmState, StateWithAsm } from '../asm-state';
 import { getAsmState } from './feature.selector';
