+++ feature-libs/asm/occ/adapters/occ-asm.adapter.spec.ts
@@ -6,12 +6,14 @@ import {
 import { TestBed } from '@angular/core/testing';
 import {
   AsmConfig,
-  CustomerSearchOptions,
-  CustomerSearchPage,
   CUSTOMER_LISTS_NORMALIZER,
   CUSTOMER_SEARCH_PAGE_NORMALIZER,
 } from '@spartacus/asm/core';
-import { CustomerListsPage } from '@spartacus/asm/root';
+import {
+  CustomerListsPage,
+  CustomerSearchOptions,
+  CustomerSearchPage,
+} from '@spartacus/asm/root';
 import {
   BaseOccUrlProperties,
   BaseSiteService,
