+++ feature-libs/asm/occ/adapters/occ-asm.adapter.ts
@@ -9,12 +9,15 @@ import { Injectable } from '@angular/core';
 import {
   AsmAdapter,
   AsmConfig,
-  CustomerSearchOptions,
-  CustomerSearchPage,
   CUSTOMER_LISTS_NORMALIZER,
   CUSTOMER_SEARCH_PAGE_NORMALIZER,
 } from '@spartacus/asm/core';
-import { BindCartParams, CustomerListsPage } from '@spartacus/asm/root';
+import {
+  BindCartParams,
+  CustomerListsPage,
+  CustomerSearchOptions,
+  CustomerSearchPage,
+} from '@spartacus/asm/root';
 import {
   BaseSiteService,
   ConverterService,
