+++ feature-libs/asm/root/facade/asm-customer-list.facade.ts
@@ -8,6 +8,7 @@ import { Injectable } from '@angular/core';
 import { facadeFactory, QueryState } from '@spartacus/core';
 import { Observable } from 'rxjs';
 import { ASM_FEATURE } from '../feature-name';
+import { CustomerSearchOptions, CustomerSearchPage } from '../model/asm.models';
 import { CustomerListsPage } from '../model/customer-list.model';
 
 @Injectable({
@@ -22,4 +23,8 @@ import { CustomerListsPage } from '../model/customer-list.model';
 export abstract class AsmCustomerListFacade {
   abstract getCustomerLists(): Observable<CustomerListsPage | undefined>;
   abstract getCustomerListsState(): Observable<QueryState<CustomerListsPage>>;
+  abstract customerListCustomersSearch(options: CustomerSearchOptions): void;
+  abstract getCustomerListCustomersSearchResults(): Observable<CustomerSearchPage>;
+  abstract getCustomerListCustomersSearchResultsLoading(): Observable<boolean>;
+  abstract customerListCustomersSearchReset(): void;
 }
