+++ feature-libs/asm/root/model/index.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export * from './asm.models';
 export * from './augmented-core.model';
 export * from './cart-binding.models';
 export * from './customer-list.model';
